refactor(vfs/syscall)：迁移ftruncate
